+++ core/src/main/java/com/epam/esm/mapper/GiftCertificateRowMapper.java
@@ -16,7 +16,7 @@ public class GiftCertificateRowMapper implements RowMapper<GiftCertificate> {
     public GiftCertificate mapRow(ResultSet rs, int rowNum) throws SQLException {
         GiftCertificate giftCertificate = new GiftCertificate();
 
-        giftCertificate.setId(rs.getLong("id"));
+        giftCertificate.setId(rs.getInt("id"));
         giftCertificate.setName(rs.getString("name"));
         giftCertificate.setDescription(rs.getString("description"));
         giftCertificate.setPrice(rs.getInt("price"));
